+++ jina/jaml/parsers/flow/legacy.py
@@ -1,16 +1,17 @@
-from typing import Dict, Any
+from typing import Dict, Any, Type
 
 from ..base import VersionedYAMLParser
-from ....flow import Flow
+from ....flow.base import BaseFlow
 from ....helper import expand_env_var
 
 
 class LegacyParser(VersionedYAMLParser):
     version = 'legacy'  # the version number this parser designed for
 
-    def parse(self, cls: type, data: Dict) -> 'Flow':
+    def parse(self, cls: Type['BaseFlow'], data: Dict) -> 'BaseFlow':
         """Return the Flow YAML parser given the syntax version number
 
+        :param cls: target class type to parse into, must be a :class:`JAMLCompatible` type
         :param data: flow yaml file loaded as python dict
         """
         p = data.get('with', {})  # type: Dict[str, Any]
@@ -30,7 +31,7 @@ class LegacyParser(VersionedYAMLParser):
 
         return obj
 
-    def dump(self, data: 'Flow') -> Dict:
+    def dump(self, data: 'BaseFlow') -> Dict:
         """Return the dictionary given a versioned flow object
 
         :param data: versioned flow object
